mainloop: do not register signal handlers as exclusive
As long as we have a single signal registrations, these will be executed
anyway. IV_SIGNAL_FLAG_EXCLUSIVE are for registrations that want
to suppress previous registrations.

I intend to use SIGINT in the debugger to break into the debugger to
allow executing commands even outside of the pipeline (e.g. setting
breakpoints for the future).

Signed-off-by: Balazs Scheidler <balazs.scheidler@axoflow.com>